feat(api): Remove controllers and routes
Remove getNote, updateNote, and deleteNote controllers and routers,
given this would be a customer only app, it doesn't make sense to give
the user control to edit or remove notes, since only admins should do
that on their end. So will leave only the addNote and getNotes
controllers and routes.